build login screen logic issues resolved
